+++ Amplify/Categories/DataStore/Internal/DataStoreCategory+Configurable.swift
@@ -24,10 +24,7 @@ extension DataStoreCategory: CategoryConfigurable {
             throw error
         }
 
-        for (pluginKey, pluginConfiguration) in configuration.plugins {
-            let plugin = try getPlugin(for: pluginKey)
-            try plugin.configure(using: pluginConfiguration)
-        }
+        try Amplify.configure(plugins: Array(plugins.values), using: configuration)
 
         isConfigured = true
     }