+++ AmplifyPlugins/Analytics/AWSPinpointAnalyticsPlugin/AWSPinpointAnalyticsPlugin+Configure.swift
@@ -18,7 +18,7 @@ extension AWSPinpointAnalyticsPlugin {
     /// - Parameter configuration: The configuration specified for this plugin
     /// - Throws:
     ///   - PluginError.pluginConfigurationError: If one of the configuration values is invalid or empty
-    public func configure(using configuration: Any) throws {
+    public func configure(using configuration: Any?) throws {
         guard let config = configuration as? JSONValue else {
             throw PluginError.pluginConfigurationError(
                 AnalyticsPluginErrorConstant.decodeConfigurationError.errorDescription,