+++ AmplifyPlugins/Predictions/AWSPredictionsPluginTests/ConfigurationTests/PredictionsPluginConfigurationTests.swift
@@ -180,4 +180,21 @@ class PredictionsPluginConfigurationTests: XCTestCase {
         }
     }
 
+    func testThrowsOnMissingConfig() throws {
+        let plugin = AWSPredictionsPlugin()
+        try Amplify.add(plugin: plugin)
+
+        let categoryConfig = PredictionsCategoryConfiguration(plugins: ["NonExistentPlugin": true])
+        let amplifyConfig = AmplifyConfiguration(predictions: categoryConfig)
+        do {
+            try Amplify.configure(amplifyConfig)
+            XCTFail("Should have thrown a pluginConfigurationError if not supplied with a plugin-specific config.")
+        } catch {
+            guard case PluginError.pluginConfigurationError = error else {
+                XCTFail("Should have thrown a pluginConfigurationError if not supplied with a plugin-specific config.")
+                return
+            }
+        }
+    }
+
 }