+++ AmplifyTestCommon/Mocks/MockAnalyticsCategoryPlugin.swift
@@ -12,7 +12,7 @@ class MockAnalyticsCategoryPlugin: MessageReporter, AnalyticsCategoryPlugin {
         return "MockAnalyticsCategoryPlugin"
     }
 
-    func configure(using configuration: Any) throws {
+    func configure(using configuration: Any?) throws {
         notify()
     }