+++ AmplifyTestCommon/Mocks/MockDataStoreCategoryPlugin.swift
@@ -14,7 +14,7 @@ class MockDataStoreCategoryPlugin: MessageReporter, DataStoreCategoryPlugin {
         return "MockDataStoreCategoryPlugin"
     }
 
-    func configure(using configuration: Any) throws {
+    func configure(using configuration: Any?) throws {
         notify()
     }